Only delete IssuedCoupon if it is reserved (Sepulsa)
Only update availability if the IssuedCoupon deleted (Sepulsa) or status reverted to available (Hot Deals).

+++ app/models/PaymentTransaction.php
@@ -204,9 +204,21 @@ class PaymentTransaction extends Eloquent
 
         // If it is Sepulsa, then remove the IssuedCoupon record.
         if ($this->forSepulsa()) {
-            Log::info('Payment: Transaction ID ' . $this->payment_transaction_id . '. Removing reserved sepulsa voucher.');
+            // TODO: Check if the coupon is already issued. If so, then what should we do?
+            $issuedCoupon = IssuedCoupon::where('transaction_id', $this->payment_transaction_id)->first();
 
-            IssuedCoupon::where('transaction_id', $this->payment_transaction_id)->delete();
+            if ($issuedCoupon->status === IssuedCoupon::STATUS_RESERVED) {
+                Log::info('Payment: Transaction ID ' . $this->payment_transaction_id . '. Removing reserved sepulsa voucher.');
+                IssuedCoupon::where('transaction_id', $this->payment_transaction_id)->delete();
+
+                // Update the availability...
+                if (! empty($this->coupon)) {
+                    $this->coupon->updateAvailability();
+                }
+            }
+            else {
+                Log::info('Payment: Transaction ID ' . $this->payment_transaction_id . '. Voucher is already issued. Do NOTHING at the moment.');
+            }
         }
         // If it is Hot Deals, then reset the IssuedCoupon state.
         else if ($this->forHotDeals()) {
@@ -219,13 +231,14 @@ class PaymentTransaction extends Eloquent
 
             if (! empty($issuedCoupon)) {
                 $issuedCoupon->makeAvailable();
+
+                // Update the availability...
+                if (! empty($this->coupon)) {
+                    $this->coupon->updateAvailability();
+                }
+
                 Log::info('Payment: hot deals coupon reverted. IssuedCoupon ID: ' . $issuedCoupon->issued_coupon_id);
             }
-        }
-
-        // Update the availability...
-        if (! empty($this->coupon)) {
-            $this->coupon->updateAvailability();
         }
     }
 }
